Split runner project policy tests
